remove install tools
